\name{cross.beds1}
\alias{cross.beds1}
\docType{data}
\title{ Cross-beds azimuths (I)
}
\description{
This dataset corresponds to azimuths of cross-beds in the Kamthi river (India). Originally analized by SenGupta and Rao (1966) and included in Table 1.5 in 
Mardia (1972), the dataset collects 580 azimuths of layers lying oblique to principal accumulation surface along the river, being these layers known as 
cross-beds.
}
\usage{data(cross.beds1)}
\format{
  A single-column data frame with 580 observations in radians.
}
\details{
Data were originally recorded in degrees.
}
\source{
Mardia, K.V. (1972) \emph{Statistics of Directional Data}. Academic Press, New York.

SenGupta, S. and Rao, J.S. (1966) Statistical analysis of cross--bedding azimuths from the Kamthi formation around Bheemaram, 
Pranhita: Godavari Valley. \emph{Sankhya: The Indian Journal of Statistics, Series B}, \bold{28}, 165--174.
}
\examples{
data(cross.beds1)
}
\keyword{datasets}

