\name{cross.beds2}
\alias{cross.beds2}
\docType{data}
\title{ Cross-beds (II)
}
\description{
A dataset of cross-beds measurements from Himalayan molasse in Pakistan presented in Fisher (1993). This dataset collects 104 measurements of Chaudan 
Zam large bedforms.
}
\usage{data(cross.beds2)}
\format{
 A single-column data frame with 104 observations in radians.
}
\details{
Data were originally recorded in degrees.
}
\source{
Fisher, N.I. (1993) \emph{Statistical Analysis of Circular Data}. Cambridge University Press, Cambridge, U.K.
}
\examples{
data(cross.beds2)
}
\keyword{datasets}

