\name{periwinkles}
\alias{periwinkles}
\docType{data}
\title{ Orientations of dragonflies
}
\description{
These data, presented in Fisher and Lee (1992), contain distance and directions of movements from small blue periwinkles after relocation.
}
\usage{data(periwinkles)}
\format{
  A two-column data frame with 73 observations. Distances are measured in centimeters and directions are measured in degrees.
}
\source{
Fisher, N. I. and Lee, A. J. (1992) Regression models for angular responses. \emph{Biometrics}, \bold{48}, 665--677.
}
\examples{
data(periwinkles)
}
\keyword{datasets}

