
\name{NPCirc-internal}
\alias{amisefr}
\alias{amisefr2}
\alias{amsephi}
\alias{aux_CRSC}
\alias{aux_cross_validation}
\alias{aux_refined}
\alias{bw.dpi}
\alias{bw.ste}
\alias{cosmoments}
\alias{cross_validation}
\alias{CRSC}
\alias{d_fun}
\alias{dcirculardr}
\alias{ddvonmises}
\alias{dvonmisesd0}
\alias{dvonmisesd1}
\alias{dvonmisesd2}
\alias{dvonmisesd3}
\alias{dvonmisesd4}
\alias{dvonmisesd5}
\alias{dvonmisesd6}
\alias{dvonmisesd8}
\alias{DensCircRad}
\alias{dmixdvonmises}
\alias{est_bias}
\alias{est_variance}
\alias{estimation_gamma}
\alias{intmixddvonmises}
\alias{gammabw}
\alias{intP1}
\alias{intP2}
\alias{int.Simp}
\alias{kernCL}
\alias{loc_loglik}
\alias{loglik_cv}
\alias{matrixK_block}
\alias{MISECC}
\alias{MISECL}
\alias{MISELC}
\alias{mle.mixvm}
\alias{my_fun_DoublePois}
\alias{my_fun_loglik_cv_nu}
\alias{NN.circ}
\alias{NN}
\alias{Omega_a}
\alias{Param_DP}
\alias{pois_p1}
\alias{R_CV_grid_modereg_CircCirc}
\alias{R_CV_grid_modereg_CircLin}
\alias{R_CV_grid_modereg_LinCirc}
\alias{R_CV_modereg_CircCirc}
\alias{R_CV_modereg_CircLin2}
\alias{R_CV_modereg_LinCirc}
\alias{R_EM_vMN_select}
\alias{R_EM_vMvM_select}
\alias{R_cv_re_modereg_CircCirc}
\alias{R_cv_re_modereg_LinCirc}
\alias{R_modereg_CircCirc}
\alias{R_modereg_CircLin}
\alias{R_modereg_LinCirc}
\alias{R_myfun_DPoisG}
\alias{R_myfun_DPoisM}
\alias{R_re_CV_modereg_CircLin2}
\alias{refined_rule}
\alias{RegCircCirc}
\alias{RegCircLin}
\alias{RegLinCirc}
\title{Internal NPCirc Functions}
\description{
 Internal NPCirc functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
